Move base_command out of the commands dir
The base_command has been moved out of the command dir as it itself
is not a command. This is to allow easier autocomplete
