+++ src/base_command.rb
