+++ src/commands/hosts.rb
@@ -1,5 +1,5 @@
 
-require 'commands/base_command'
+require 'base_command'
 require 'constants'
 require 'iterator'
 require 'nodes'
