+++ src/commands/repo/use.rb
@@ -1,5 +1,5 @@
 
-require 'commands/base_command'
+require 'base_command'
 require 'rugged'
 require 'fileutils'
